MAINT: Refactored out the structured solvers testing utils
