+++ tests/structured_solvers_utils.py
